# openapi/openapi.json
{
  "info": {
    "description": "REST API for managing technical processes, products, documents, and resources in the PSS application",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "Proprietary"
    },
    "title": "Tech Process Viewer API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:7239",
      "description": "Backend PSS REST API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-APL-SessionKey",
        "description": "Session key obtained from /api/connect endpoint"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "status": {
            "type": "string",
            "description": "Error name"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "description": "Errors",
            "additionalProperties": {}
          }
        }
      },
...
      "Error1": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error description"
          },
          "status_code": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": {}
          }
        },
        "required": [
          "error"
        ]
      },
...
      "BusinessProcessType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Process type ID"
          },
          "name": {
            "type": "string",
            "description": "Process type name"
          }
        },
        "required": [
          "id"
        ]
      },
      "BusinessProcessElement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Element system ID"
          },
          "type": {
            "type": "string",
            "description": "Element type"
          }
        },
        "required": [
          "id"
        ]
      },
      "BusinessProcessResourceReference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Resource system ID"
          },
          "type": {
            "type": "string",
            "description": "Resource type"
          }
        },
        "required": [
          "id"
        ]
      },
      "BusinessProcessAttributes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Business process ID (business key)"
          },
          "name": {
            "type": "string",
            "description": "Process name"
          },
          "description": {
            "type": "string",
            "default": "",
            "description": "Process description"
          },
          "customized": {
            "type": "boolean",
            "description": "Whether process is customized"
          },
          "type": {
            "description": "Process type reference",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessProcessType"
              }
            ]
          },
          "elements": {
            "type": "array",
            "description": "Sub-processes/phases",
            "items": {
              "$ref": "#/components/schemas/BusinessProcessElement"
            }
          },
          "resources": {
            "type": "array",
            "description": "Process resources",
            "items": {
              "$ref": "#/components/schemas/BusinessProcessResourceReference"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "BusinessProcess": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "System ID"
          },
          "type": {
            "type": "string",
            "description": "Entity type"
          },
          "access": {
            "type": "string",
            "description": "Access level"
          },
          "attributes": {
            "$ref": "#/components/schemas/BusinessProcessAttributes"
          }
        },
        "required": [
          "attributes",
          "id",
          "type"
        ]
      },
      "BusinessProcessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          },
          "data": {
            "description": "Business process data",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessProcess"
              }
            ]
          },
          "process_id": {
            "type": "integer",
            "description": "Created/updated process ID"
          },
          "resource_id": {
            "type": "integer",
            "description": "Resource ID (returned from resource-add endpoints)",
            "nullable": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "BusinessProcessResourceCreate": {
        "type": "object",
        "properties": {
          "type_id": {
            "type": "integer",
            "description": "Resource type ID"
          },
          "object_id": {
            "type": "integer",
            "description": "Object ID (organization, product, etc.)"
          },
          "value_component": {
            "type": "number",
            "description": "Resource value (e.g., time in hours)"
          },
          "unit_id": {
            "type": "integer",
            "description": "Unit of measurement ID"
          }
        },
        "required": [
          "type_id"
        ]
      },
      "BusinessProcessResourceUpdate": {
        "type": "object",
        "properties": {
          "value_component": {
            "type": "number",
            "description": "Resource value"
          },
          "object_id": {
            "type": "integer",
            "description": "Object ID"
          },
          "unit_id": {
            "type": "integer",
            "description": "Unit ID"
          }
        }
      },
...
      "Reference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Object system ID"
          },
          "type": {
            "type": "string",
            "description": "Object type"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "ResourceListItem": {
        "type": "object",
        "properties": {
          "resource_id": {
            "type": "integer",
            "description": "Resource system ID"
          },
          "process_id": {
            "type": "integer",
            "description": "Business process ID",
            "nullable": true
          },
          "type_name": {
            "type": "string",
            "description": "Resource type name",
            "nullable": true
          },
          "value_component": {
            "type": "number",
            "description": "Quantity/value",
            "nullable": true
          },
          "unit_name": {
            "type": "string",
            "description": "Unit name",
            "nullable": true
          }
        }
      },
      "ResourceListResponse": {
        "type": "object",
        "properties": {
          "resources": {
            "type": "array",
            "description": "List of resources",
            "items": {
              "$ref": "#/components/schemas/ResourceListItem"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of resources"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "per_page": {
            "type": "integer",
            "description": "Items per page"
          }
        }
      },
      "ResourceCreate": {
        "type": "object",
        "properties": {
          "process_id": {
            "type": "integer",
            "description": "Business process system ID"
          },
          "type_id": {
            "type": "integer",
            "description": "Resource type system ID"
          },
          "object_id": {
            "type": "integer",
            "description": "Resource object system ID",
            "nullable": true
          },
          "value_component": {
            "type": "number",
            "description": "Resource quantity/value",
            "nullable": true
          },
          "unit_id": {
            "type": "integer",
            "description": "Unit system ID",
            "nullable": true
          }
        },
        "required": [
          "process_id",
          "type_id"
        ]
      },
      "ResourceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Resource data",
            "additionalProperties": {}
          },
          "resource_id": {
            "type": "integer",
            "description": "Resource system ID"
          }
        },
        "required": [
          "success"
        ]
      },
      "Resource": {
        "type": "object",
        "properties": {
          "value_component": {
            "type": "number",
            "description": "Resource quantity/value",
            "nullable": true
          },
          "unit_component": {
            "description": "Unit reference",
            "allOf": [
              {
                "$ref": "#/components/schemas/Reference"
              }
            ],
            "nullable": true
          },
          "object": {
            "description": "Resource object reference",
            "allOf": [
              {
                "$ref": "#/components/schemas/Reference"
              }
            ],
            "nullable": true
          },
          "type": {
            "description": "Resource type reference",
            "allOf": [
              {
                "$ref": "#/components/schemas/Reference"
              }
            ],
            "nullable": true
          },
          "process": {
            "description": "Business process reference",
            "allOf": [
              {
                "$ref": "#/components/schemas/Reference"
              }
            ],
            "nullable": true
          }
        }
      },
      "ResourceDetail": {
        "type": "object",
        "properties": {
          "sys_id": {
            "type": "integer",
            "description": "System ID"
          },
          "type": {
            "type": "string",
            "description": "Entity type"
          },
          "access": {
            "type": "string",
            "description": "Access level"
          },
          "attributes": {
            "description": "Resource attributes",
            "allOf": [
              {
                "$ref": "#/components/schemas/Resource"
              }
            ]
          }
        }
      },
      "ResourceUpdate": {
        "type": "object",
        "properties": {
          "value_component": {
            "type": "number",
            "description": "Resource quantity/value",
            "nullable": true
          },
          "unit_id": {
            "type": "integer",
            "description": "Unit system ID",
            "nullable": true
          },
          "object_id": {
            "type": "integer",
            "description": "Resource object system ID",
            "nullable": true
          }
        }
      },
      "ResourceType": {
        "type": "object",
        "properties": {
          "sys_id": {
            "type": "integer",
            "description": "apl_business_process_resource_type sys_id (use for create_resource.type_id)"
          },
          "id": {
            "type": "string",
            "description": "Resource type ID"
          },
          "name": {
            "type": "string",
            "description": "Resource type name"
          },
          "description": {
            "type": "string",
            "description": "Description",
            "nullable": true
          }
        }
      },
      "ResourceTypeListResponse": {
        "type": "object",
        "properties": {
          "resource_types": {
            "type": "array",
            "description": "List of resource types",
            "items": {
              "$ref": "#/components/schemas/ResourceType"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of types"
          }
        }
      },
...
    },
    "responses": {
      "UNPROCESSABLE_ENTITY": {
        "description": "Unprocessable Entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "DEFAULT_ERROR": {
        "description": "Default error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "SessionKey": []
    }
  ],
  "paths": {
...
    "/api/v1/business-processes/{process_id}/resources": {
      "post": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Business process not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProcessResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessProcessResourceCreate"
              }
            }
          }
        },
        "summary": "Add resource to business process",
        "tags": [
          "business_processes"
        ],
        "description": "Add a resource to a business process"
      },
      "parameters": [
        {
          "in": "path",
          "name": "process_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
    "/api/v1/business-processes/{process_id}/resources/{resource_id}": {
      "put": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProcessResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessProcessResourceUpdate"
              }
            }
          }
        },
        "summary": "Update business process resource",
        "tags": [
          "business_processes"
        ],
        "description": "Update a business process resource"
      },
      "delete": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Delete business process resource",
        "tags": [
          "business_processes"
        ],
        "description": "Delete a business process resource"
      },
      "parameters": [
        {
          "in": "path",
          "name": "process_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        },
        {
          "in": "path",
          "name": "resource_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
...
    "/api/v1/resources/": {
      "get": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "List all resources",
        "tags": [
          "resources"
        ],
        "description": "Get list of resources with optional filtering"
      },
      "post": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceCreate"
              }
            }
          }
        },
        "summary": "Create a new resource",
        "tags": [
          "resources"
        ],
        "description": "Create a new resource"
      }
    },
    "/api/v1/resources/{resource_id}": {
      "get": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDetail"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Get resource by ID",
        "tags": [
          "resources"
        ],
        "description": "Get resource details by ID"
      },
      "put": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceUpdate"
              }
            }
          }
        },
        "summary": "Update resource",
        "tags": [
          "resources"
        ],
        "description": "Update resource"
      },
      "delete": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Delete resource",
        "tags": [
          "resources"
        ],
        "description": "Delete resource"
      },
      "parameters": [
        {
          "in": "path",
          "name": "resource_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
    "/api/v1/resources/types": {
      "get": {
        "responses": {
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceTypeListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "List all resource types",
        "tags": [
          "resources"
        ],
        "description": "Get list of resource types"
      }
    },
...
  },
...
  "openapi": "3.0.3"
}